Implemented coutdown timer to start of presentation
This implementation should be considered a hack for now. It has to be
refactored in the future. For example if you start your presentation timer
early by pressing one of the navigation keys during a countdown is running you
are not able to go back into the countdown mode by pressing POS1 for example.