Overview: Center icons in available space
This is accomplished by making Overview into a Gtk.Alignment that
contains a ScrolledWindow.  The ScrolledWindow makes a size request to
be just big enough to hold all of the icons, and the Alignment manages
the space around it to center it.

For some reason, the vertical scrollbar of the ScrolledWindow is
triggered whether it is needed or not.  So we set its policy to NEVER
when we're sure we have enough room and to AUTOMATIC when we think we
don't.