Overview: Enable better mouse control
Mouse motion now sets the selection and a left click goes back to the
main view.  Actually, it's the button release that does this.  This
means if you click and drag, you'll be taken to the slide where you
release the button.  I don't know if this is desired or not.