Overview: Don't redo the structure when changing overlays
When changing the overlay settings in the overview, we will only be
removing slides.  Therefore, we don't need to redo the structure and
re-render all of the icons -- we can just cut that slide out off the
ListStore.  We might get a slightly sub-optimal arrangement, but that's
better than rearranging everything.  The next time the overview is
hidden, we'll get it set up all pretty again.