ControlledMovie: Add capability filter
Place a capsfilter upstream of the overlay element.  Here, we set the
final size of the movie, so that the cairo drawing comes out nice and
crisp.  Also, we require a frame rate of at least 25 fps to keep the
controls reasonably snappy.

Note that the videorate element causes a hitch in the playback of one
test file.  This is likely a problem with the file, but we may want to
insert the videorate only when necessary.