Made the strict_views slightly smaller
It seems that taking the whole height of the presenter window can
trigger problems under certain situations (pdfpc started on a monitor
with a bigger resolution than the target monitor of the presentation
window). I do not know for sure why this happens, but this seems to
solve it for the time being.

Also activated double buffering for the views, as it seems to reduce
some effects (substituting them with a gray background).

The view and renderer architecture have to be reconsidered...