Reconfiguration of link handling from signal-based to actions
What started as an effort to get the movies created earlier turned into
a rather massive rewrite.  Now, when a page is loaded, an action is
created for each link (and soon other things) found.  This action is
responsible for handling mouse-in, mouse-out, and button-click events on
itself.  Something like this will be needed for auto-play movies, for
example, and it seemed better to keep the links and movies using the
same mechanism.