Actions: Separate constructors and initializers
This allows us to create actions with Object.new(type) and still
initialize them separately.  Since we generally create actions with the
new_from... methods, this doesn't make life more difficult.  It also
lets us get rid of the blank constructors which were sort of silly.

In the movie classes, there are init_other methods that allow one object
to initialize another.  These are needed because I couldn't figure out
how to cast the newly created object to the existing object's type so
that the correct initializer would be called.