Be more gentle when saving metadata
Previously the .pdfpc file was deleted if pdfpc deemed it unnecessary and
was replaced completely with a new one if not. This has some disadvantages:
 - no links can be used in place of .pdfpc files (neither symlinks nor hard links)
 - file permissions get overwritten every time

This patch uses File.replace_contents() instead of FileUtils.set_contents() to
modify the metadata file. It does so if there is some actual useful metadata
to write or if threre was a metadata file before. This will work fine with links
but does no longer delete unnecessary metadata files after they were
created already. This should be no problem because usually metadata is
added not removed and the redundant file is harmless anyway.