Support playing subsequence of movies via ?start=xxx&stop=xxx
When movies are embedded in the PDF via a link (ie
\href{run:movie.avi}{...}), pdfpc will now treat queries arguments
'start' and 'stop' as indication of a start and stop time, in seconds
from the beginning of the movie.

Those can be used in combination with existing 'autostart' and 'loop'.

However, a remaining (mysterious) bug prevent 'loop' of looping when
using 'stop' ('start' works)

The seek bar highlights the sub-sequence and still let the user to
manually seek out of it.