Switch to GStreamer 1.0
For some reason, now we treat the XWindow IDs as uint* instead of uint.
I don't know if it was wrong before, or why it worked if that was so.

This isn't working quite right when the movie's aspect ratio doesn't
match the area in which it is being displayed.