Use CSS for styling, rather than setting widget properties
Instead of setting the cursor in the progress entry to be the same
color as the background, we make it go away completely via the
sensitive property.

Note that the progress bar text in in the progress bar for Gtk < 3.14
and above the bar for newer versions, so we have to choose a color that
will work in both configurations.