Use icon view key bindings in overview
There are two parts to making this work.  First the window level
keypress handlers trap events before the widgets can get at them.  When
the page count or notes widget has focus, we block the window-level
handler completely, but we don't want to do that for the overview.  So
instead we pass the key event through to the icon view manually.

Our custom handlers on the overview will run always, but the built-in
ones will only run if the icon view has focus.  This in turn means that
its top-level window must have focus to.  So we make sure that happens
when we switch to overview mode.  A user could mess this up by switching
focus to the presentation window while in overview mode, but that's not
a likely action.