Set the CWD to be the directory of the .pdfpc file
More specifically, set the current directory to be the directory
containing the argument to the script.  This may be the .pdfpc file
itself, or it may be a .pdf file.  In the latter case, we'll search for
the .pdfpc file in that directory, so it comes out the same way.

The point of this is to ensure that we can write the path the .pdf file
from the current directory into the .pdfpc file and have everything
work.  Closes #94.