Associate the last PDF page of a slide with the slide, not the first
This change is an alternative to https://github.com/pdfpc/pdfpc/pull/141

Whenever a lookup from user slide to real slide is performed, return the
last PDF page associated with this user slide number, instead of the
first one. This results in overlay navigation jumping to the completely
revealed slide instead of to the first page of a slide. Furthermore,
this results in the preview window showing the fully revealed current
slide rather than the next one. This also results in the fully revealed
slides being visibile in the overview screen.