Respect GDK_SCALE factor
If GDK_SCALE is used (in case of HiDPI), GDK divides every widget size
by GDK_SCALE. Later, it scales it up to the original size of the
monitor.

This results in poorly rendered slides.

We increase the Renderer.Pdf size by GDK_SCALE and shrink it for
displaying it to the user.

Use a Cairo CellRenderer (instead of pixbuf based) for drawing the
slides on the overview.

Icons look nice as well (use cairo surfaces)