Detect the GDK backend
Wayland and X11 works different when it comes to fullscreening and
window moving.

Wayland has no concept of global window positions (just positions
relative to the display). Moving, fullscreening and the checks if
everything worked do not work in Wayland sessions.

There is a API Gtk.Window.fullscreen_on_monitor which works on Wayland
but do nothing on X11.

To have different code paths for X11/Wayland, we need to know the GDK
backend at runtime. GDK has some C macros to do the checks, but we
can't access them in vala directly. To avoid this problem, we have
two small helper functions, written in C, which are called from vala.