Changed overview layout
On Fedora 25 with wayland, the sizing of the overview window did not
worked. This resulted in a too heigh overview window (thus moving the
bottom_row out of the screen).

Changed the way the overview size is set and changed scrolledwindow
policy to always use scroll bars. Without this "always scroll" option
the size of the inner widget of the scrolledwindow resulted in the out
of screen bug.