Reworked bottom bar
Reduced the minimum size of progressbar and gtkentry (the slide
counter).

On small screen resolutions or in some GDK_SCALE settings both widgets
became too large and the presenter window became larger than the
screen size.

This resulted in either a overlap with the presentation view or in
unfullscreening in some WMs (like ubuntu).

The minimum size of progressbar is now 1px (before it was 150px). The
progressbar now expands better in the previous described situations.