Add hiding of presentation window
When 'h' key is pressed the visibility of the presentation window is
toggled.

This functionality allows me to interrupt the presentation, show live
examples and quickly continue the presentation. Without this, I had to
quit pdfpc, show the example and than start pdfpc again and manually
navigate to the last slide.