Add options --current-width (-u) and --current-height (-U)
These options are useful to make the current slide bigger (or smaller)
on the presenter screen. There was an option --current-size in pdfpc
3.1.1 that I missed in the latest version, so I am adding it back,
expanding it to be able to also set the maximum height (the wired in
default for that is 80).

I found that I need to use something like `-u 65 -U 90` to get the
presenter view I want.