show the previous user slide on the auxiliary screen
Make the auxiliary screen show the previous user slide to support the
most typical use case: use an extra beamer to show the "previous page",
i.e. what disappeared from the audience's view when the presenter
switched to the next (user) slide. This should not (and does not) require
any extra configuratin in the .pdfpc sidecar file.