Reworked pdfpc latex package for CTAN submission
The CTAN people proposed to restructure our code. The main change is
that we no longer use .dtx files to combine code and documentation (to
split it directly afterswards).

The package code belongs to pdfpc.sty file, the documentation belongs
to pdfpc-doc.tex (and will be compiled to pdf for CTAN submission)