Hide pen/eraser after 2 secs of inactivity
In versions before 872c905 ("Ignore the pointer_hidden flag in the
drawing mode", 2020-05-09), it could happen, that a drawing mode tool
was hidden and the user did not see where they will draw. The (now
reverted) commit 872c905 fixed this problem by never hiding the
drawing tool. This resulted in the drawing tool being visible in the
slide area even if the cursor was moved out of the slide area.

This commit fixes the problem differently. The drawing tool is hidden
the same way as the pointer - after two seconds of inactivity or when
the cursor is out of slide area.