Always update the pointer position, even if it is not shown.
Currently, when the pointer is first made visible, it is drawn at (0, 0)
until the mouse is moved for the first time. Similarly, later, the pointer
is always shown at its previous location until the mouse is moved. This
behaviour causes the pointer to jump around more than necessary.

This change always updates the pointer position even when it is not shown,
so it always appears at the mouse position.