Change demo zip file for solving #740
It seems that the behaviour of asset upload/download changes
such that the original file name of the asset is preserved.

When attaching release assets we want a pdfpc- prefix in our file
names. Change the scripts such that we have this prefix for video zip
in the first place like in pdfpc-demo.pdf.